+++ packages/datadog_flutter_plugin/datadog_flutter_plugin/test/rum/inv_metric_provider_test.dart
@@ -3,8 +3,8 @@
 // Copyright 2025-Present Datadog, Inc.
 
 import 'package:datadog_common_test/datadog_common_test.dart';
-import 'package:datadog_flutter_plugin/datadog_flutter_plugin.dart';
 import 'package:datadog_flutter_plugin/src/rum/inv_metric_provider.dart';
+import 'package:datadog_flutter_plugin/src/rum/rum.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 const defaultFirstBuildTime = Duration(milliseconds: 38);